smart_city/app_smart/api/viewsets.py
from django.contrib.auth.models import User
from rest_framework import generics, permissions
from app_smart.api import serializers #app_smart é o app criado.
from rest_framework.response import Response
from rest_framework import status
from ..models import Sensor #.. é para voltar dois níveis para encontrar o
 # arquivo models.py
from rest_framework import viewsets
from app_smart.api.filters import SensorFilter
from django_filters.rest_framework import DjangoFilterBackend
import os
import csv
from rest_framework.views import APIView
from rest_framework.response import Response
from app_smart.models import Sensor 


class CreateUserAPIViewSet(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = serializers.UserSerializer
    permission_classes = [permissions.IsAdminUser] 

    def post(self, request, *args, **kwargs):
        return self.create(request, *args, **kwargs)
    
class SensorViewSet(viewsets.ModelViewSet):
    queryset = Sensor.objects.all()
    serializer_class = serializers.SensorSerializer
    permission_classes = [permissions.IsAuthenticated]
    filter_backends = [DjangoFilterBackend]
    filterset_class = SensorFilter

